Databricks_Demos/smart_claims_analysis/requirements.txt
Deprecated==1.2.18
Flask==3.0.3
GitPython==3.1.44
Jinja2==3.1.6
MarkupSafe==2.1.5
PyYAML==6.0.2
Pygments==2.19.1
Werkzeug==3.0.6
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash==2.18.1
dash_ag_grid==31.2.0
dash_mantine_components==0.14.4
databricks-sdk==0.33.0
databricks-sql-connector==3.4.0
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.115.0
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
google-auth==2.39.0
gradio==4.44.0
gradio_client==1.3.0
gunicorn==23.0.0
h11==0.16.0
htmltools==0.6.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
jiter==0.9.0
jsonschema-specifications==2025.4.1
jsonschema==4.23.0
kiwisolver==1.4.8
linkify-it-py==2.0.3
lz4==4.4.4
markdown-it-py==3.0.0
matplotlib==3.10.1
mdit-py-plugins==0.4.2
mdurl==0.1.2
mlflow-skinny==2.16.2
narwhals==1.37.1
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.78.1
openpyxl==3.1.5
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==10.4.0
plotly-resampler==0.10.0
plotly==5.24.1
prompt_toolkit==3.0.51
protobuf==5.29.4
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
questionary==2.1.0
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
ruff==0.11.8
semantic-version==2.10.0
shellingham==1.5.4
shiny==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.38.6
streamlit==1.38.0
tenacity==8.5.0
thrift==0.20.0
toml==0.10.2
tomlkit==0.12.0
tornado==6.4.2
tqdm==4.67.1
tsdownsample==0.1.4.1
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.4.0
uvicorn==0.30.6
uvloop==0.21.0
watchdog==4.0.2
watchfiles==1.0.5
wcwidth==0.2.13
websockets==12.0
wrapt==1.17.2
zipp==3.21.0
